Move dispute info data to own component DisputeInfoContent
